Initial commit with updates to NewItems plus creation of reusable components.
