+++ src/components/explore/ExploreItems.jsx
@@ -2,6 +2,8 @@ import React from "react";
 import { Link } from "react-router-dom";
 import AuthorImage from "../../images/author_thumbnail.jpg";
 import nftImage from "../../images/nftImage.jpg";
+import Countdown from "../reusable/Countdown";
+
 
 const ExploreItems = () => {
   return (
