+++ src/components/reusable/Countdown.jsx
@@ -1,0 +1,54 @@
+import React, { useState, useEffect } from 'react';
+
+const Countdown = ({ expiryDate }) => {
+  const [timeLeft, setTimeLeft] = useState(0);
+
+  useEffect(() => {
+    if (!expiryDate) {
+      setTimeLeft(0);
+      return;
+    }
+
+    const calculateTimeLeft = () => {
+      const now = Date.now();
+      const difference = new Date(expiryDate).getTime() - now;
+
+
+      if (difference <= 0) {
+        return 0;
+      }
+      return difference;
+    };
+
+    setTimeLeft(calculateTimeLeft());
+
+    const timer = setInterval(() => {
+      setTimeLeft(calculateTimeLeft());
+    }, 1000);
+
+    return () => clearInterval(timer);
+  }, [expiryDate]);
+
+  const getFormattedTime = () => {
+    if (timeLeft <= 0) {
+      return "Time's up!";
+    }
+
+    const totalSeconds = Math.floor(timeLeft / 1000);
+    const hours = Math.floor(totalSeconds / 3600);
+    const minutes = Math.floor((totalSeconds % 3600) / 60);
+    const seconds = totalSeconds % 60;
+
+    return `${hours.toString().padStart(2, "0")}h ${minutes
+      .toString()
+      .padStart(2, "0")}m ${seconds.toString().padStart(2, "0")}s`;
+  };
+
+  return (
+    <div className="de_countdown">
+      {getFormattedTime()}
+    </div>
+  );
+};
+
+export default Countdown;
